\docType{data}

\name{HFD}

\alias{HFD}

\title{HFD Dataset}

\description{
This dataset is dataframe of the metabolite dataset contain only chemical information. 
}

\usage{data(HFD)}

\value{
A list with the following elements in the variable \code{HFD}:
	\item{SMILES}{A vector of secondary IDs of the each metabolites .}
	\item{CHEMICAL_ID}{A vector of chemical ID number or each metabolite.}
	\item{PUBCHEM}{A vector of identifier ID number from PUBCHEM database for chemical molecules and their activities in biological assays.}
	\item{CHEMSPIDER}{A vector of a unique identifier from CHEMSPIDER database each molecule.}
	\item{HMDB}{A vector containing all information relative to the feature of the each metabolite.}
}

\examples{
 data(HFD)

}

\keyword{datasets}
