\name{write.gct}
\alias{write.gct}
\title{
Write a GCT file
}
\description{
This function write a file in the format GCT defined by GenePattern.
}
\usage{
write.gct(es, address)
}
\arguments{
  \item{es}{A matrix.}
  \item{address}{The address of the file should be saved.}
}
\value{
No return value. If an invalid address is inserted, the function will generate an error.
}
\seealso{
  \code{\link{write.gmt}}, \code{\link{write.cls}}
}

