\name{nameMet}
\alias{nameMet}
\title{
Name of metabolites
}
\description{
This function extracts the metabolite's names from the list generated by \code{\link{readMet}} function.
}
\usage{
nameMet(doc)
}
\arguments{
  \item{doc}{A list of metabolites information produced by \code{\link{readMet}} function.}
}

\value{
A data frame contains the names of each metabolite.
}



\seealso{
  \code{\link{readMet}}
}
\examples{

data(Metabolites)
nam=nameMet(Metabolites$readMet)

}