\name{write.cls}
\alias{write.cls}
\title{
Write a CLS file
}
\description{
This function write a file in the format CLS defined by GenePattern.
}
\usage{
write.cls(es, address)
}
\arguments{
  \item{es}{A matrix.}
  \item{address}{The address of the file should be saved.}
}
\value{
No return value. If an invalid address is inserted, the function will generate an error.
}
\seealso{
  \code{\link{write.gmt}}, \code{\link{write.gct}}
}

